third_party/WebKit/LayoutTests/fast/dom/shadow/shadow-disable-expected.txt

CONSOLE WARNING: line 50: The <keygen> element is deprecated and will be removed in M57, around March 2017. See https://www.chromestatus.com/features/5716060992962560 for more details.
Tests to ensure that shadow element cannot be created in elements having dynamically created shadow root.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


DIV
PASS element.createShadowRoot() is not null
SPAN
PASS element.createShadowRoot() is not null
A
PASS element.createShadowRoot() is not null
SECTION
PASS element.createShadowRoot() is not null
g
PASS element.createShadowRoot() is not null
rect
PASS element.createShadowRoot() is not null
svg
PASS element.createShadowRoot() is not null
INPUT
PASS element.createShadowRoot() threw exception InvalidStateError: Failed to execute 'createShadowRoot' on 'Element': Shadow root cannot be created on a host which already hosts an user-agent shadow tree..
EMBED
PASS element.createShadowRoot() threw exception InvalidStateError: Failed to execute 'createShadowRoot' on 'Element': Shadow root cannot be created on a host which already hosts an user-agent shadow tree..
OBJECT
PASS element.createShadowRoot() threw exception InvalidStateError: Failed to execute 'createShadowRoot' on 'Element': Shadow root cannot be created on a host which already hosts an user-agent shadow tree..
AUDIO
PASS element.createShadowRoot() threw exception InvalidStateError: Failed to execute 'createShadowRoot' on 'Element': Shadow root cannot be created on a host which already hosts an user-agent shadow tree..
VIDEO
PASS element.createShadowRoot() threw exception InvalidStateError: Failed to execute 'createShadowRoot' on 'Element': Shadow root cannot be created on a host which already hosts an user-agent shadow tree..
SELECT
PASS element.createShadowRoot() threw exception InvalidStateError: Failed to execute 'createShadowRoot' on 'Element': Shadow root cannot be created on a host which already hosts an user-agent shadow tree..
IMG
PASS element.createShadowRoot() threw exception HierarchyRequestError: Failed to execute 'createShadowRoot' on 'Element': Author-created shadow roots are disabled for this element..
TEXTAREA
PASS element.createShadowRoot() threw exception InvalidStateError: Failed to execute 'createShadowRoot' on 'Element': Shadow root cannot be created on a host which already hosts an user-agent shadow tree..
IFRAME
PASS element.createShadowRoot() threw exception HierarchyRequestError: Failed to execute 'createShadowRoot' on 'Element': Author-created shadow roots are disabled for this element..
CANVAS
PASS element.createShadowRoot() threw exception HierarchyRequestError: Failed to execute 'createShadowRoot' on 'Element': Author-created shadow roots are disabled for this element..
METER
PASS element.createShadowRoot() threw exception InvalidStateError: Failed to execute 'createShadowRoot' on 'Element': Shadow root cannot be created on a host which already hosts an user-agent shadow tree..
PROGRESS
PASS element.createShadowRoot() threw exception InvalidStateError: Failed to execute 'createShadowRoot' on 'Element': Shadow root cannot be created on a host which already hosts an user-agent shadow tree..
KEYGEN
PASS element.createShadowRoot() threw exception InvalidStateError: Failed to execute 'createShadowRoot' on 'Element': Shadow root cannot be created on a host which already hosts an user-agent shadow tree..
VIDEO
PASS element.createShadowRoot() threw exception InvalidStateError: Failed to execute 'createShadowRoot' on 'Element': Shadow root cannot be created on a host which already hosts an user-agent shadow tree..
AUDIO
PASS element.createShadowRoot() threw exception InvalidStateError: Failed to execute 'createShadowRoot' on 'Element': Shadow root cannot be created on a host which already hosts an user-agent shadow tree..
FIELDSET
PASS element.createShadowRoot() threw exception HierarchyRequestError: Failed to execute 'createShadowRoot' on 'Element': Author-created shadow roots are disabled for this element..
PASS successfullyParsed is true

TEST COMPLETE
